Reorganize "wetterdienst.ui" module namespace
This will make more headroom for other user interfaces without
cluttering the root module namespace.
