+++ docs/usage/http_api.rst
@@ -26,13 +26,13 @@ Observations
 ::
 
     # Acquire observations.
-    http localhost:7890/api/dwd/observation/values station==1048,4411 parameter==kl resolution==daily period==recent
+    http localhost:7890/api/dwd/observation/values stations==1048,4411 parameter==kl resolution==daily period==recent
 
     # Observations for specific date.
-    http localhost:7890/api/dwd/observation/values station==1048,4411 parameter==kl resolution==daily period==recent date==2020-08-01
+    http localhost:7890/api/dwd/observation/values stations==1048,4411 parameter==kl resolution==daily period==recent date==2020-08-01
 
     # Observations for date range.
-    http localhost:7890/api/dwd/observation/values station==1048,4411 parameter==kl resolution==daily period==recent date==2020-08-01/2020-08-05
+    http localhost:7890/api/dwd/observation/values stations==1048,4411 parameter==kl resolution==daily period==recent date==2020-08-01/2020-08-05
 
     # Observations with SQL.
-    http localhost:7890/api/dwd/observation/values station==1048,4411 parameter==kl resolution==daily period==recent sql=="SELECT * FROM data WHERE element='temperature_air_max_200' AND value < 2.0;"
+    http localhost:7890/api/dwd/observation/values stations==1048,4411 parameter==kl resolution==daily period==recent tidy==False sql=="SELECT * FROM data WHERE temperature_air_max_200 < 2.0;"
