+++ wetterdienst/provider/dwd/radar/README.rst
@@ -1,0 +1,6 @@
+.. _Radarprodukte: https://www.dwd.de/DE/leistungen/radarprodukte/radarprodukte.html
+.. _Radar products: https://www.dwd.de/EN/ourservices/radar_products/radar_products.html
+.. _Radarverbund: https://www.dwd.de/DE/derdwd/messnetz/atmosphaerenbeobachtung/_functions/Teasergroup/radarverbund_teaser5.html
+.. _Aktueller Radarstatus: https://www.dwd.de/DE/leistungen/radarniederschlag/rn_info/home_freie_radarstatus_kartendaten.html
+.. _Datenformat / data format - RADOLAN: https://www.dwd.de/DE/leistungen/opendata/hilfe.html#doc625266bodyText8
+.. _Datenformat / data format - HDF5: https://www.dwd.de/DE/leistungen/opendata/hilfe.html#doc625266bodyText9
