+++ wetterdienst/ui/dash/assets/styles.css
